+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `nbcl` are documented here.
 This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format,
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [UNRELEASED]
+
+### Added
+
+- Internal Object type to register data that user cant interact with.
+
 ## [0.3.4] - 2026-05-21
 
 ### Fixed